+++ lib/controllers/cron_controller.ex
@@ -1,14 +1,14 @@
 defmodule ApplicationRunner.CronController do
   use ApplicationRunner, :controller
 
-  alias ApplicationRunner.Crons.CronServices
+  alias ApplicationRunner.Crons
 
   def create(conn, %{"env_id" => env_id} = params) do
     {env_id_int, ""} = Integer.parse(env_id)
 
     with {:ok, cron} <-
            env_id_int
-           |> CronServices.create(params) do
+           |> Crons.create(params) do
       conn
       |> reply(cron)
     end
@@ -16,7 +16,7 @@ defmodule ApplicationRunner.CronController do
 
   def get(conn, %{"id" => cron_id} = _params) do
     with {:ok, cron} <-
-           CronServices.get(cron_id) do
+           Crons.get(cron_id) do
       conn
       |> reply(cron)
     end
@@ -24,17 +24,17 @@ defmodule ApplicationRunner.CronController do
 
   def all(conn, %{"env_id" => env_id, "user_id" => user_id} = _params) do
     conn
-    |> reply(CronServices.all(env_id, user_id))
+    |> reply(Crons.all(env_id, user_id))
   end
 
   def all(conn, %{"env_id" => env_id} = _params) do
     conn
-    |> reply(CronServices.all(env_id))
+    |> reply(Crons.all(env_id))
   end
 
   def update(conn, %{"id" => cron_id} = params) do
-    with {:ok, cron} <- CronServices.get(cron_id),
-         {:ok, updated_cron} <- CronServices.update(cron, params) do
+    with {:ok, cron} <- Crons.get(cron_id),
+         {:ok, updated_cron} <- Crons.update(cron, params) do
       conn
       |> reply(updated_cron)
     end
